# ddtrace/tracer: increase performance of SetTag