This change adds benchmarks for SetTag and increases the performance of
setting string values, which is our most common scenario.

```
name            old time/op    new time/op    delta
SetTagString-4     178ns ± 0%      99ns ± 0%  -44.16%

name            old alloc/op   new alloc/op   delta
SetTagString-4     16.0B ± 0%      4.0B ± 0%  -75.00%

name            old allocs/op  new allocs/op  delta
SetTagString-4      2.00 ± 0%      1.00 ± 0%  -50.00%
```